requirements.txt
requests==2.31.0
jinja2==3.1.2
python-dotenv==1.0.0
google-generativeai>=0.8.0
